add tests for deep merge
